+++ src/renderers/dom/shared/utils/voidElementTags.js
@@ -1,0 +1,24 @@
+/**
+ * Copyright 2013-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @providesModule voidElementTags
+ */
+
+'use strict';
+
+var omittedCloseTags = require('omittedCloseTags');
+
+// For HTML, certain tags cannot have children. This has the same purpose as
+// `omittedCloseTags` except that `menuitem` should still have its closing tag.
+
+var voidElementTags = {
+  menuitem: true,
+  ...omittedCloseTags,
+};
+
+module.exports = voidElementTags;
